fix: stop test assets bundled in production
- restructure ConfigFeatureFlag assets so they are no longer
exporting test code as runtime dependencies

Contributes to: strimzi/strimzi-ui#152
